We're getting very close.
We were having issues with the keys that are located in the .plist files that were merged wrong. So, for most of the night, we were fighting permission issues that weren't actually there.
We ran the podfile and got it setup with XYO.
And now, we're just working on exposing the uPort address in the Proof of Location calls.